Move handlers and styles outside the chart
